rename:함수3개를각각  파일을 새로만들어 이동,불필요 콘솔 제거 (#29)
